Allow reuse of existing keys in certificate api
Allow keys to be recycled if they have been decommissioned, by providing
an optional `public_key` and `private_key` value.

If both values are set, we can mandate the value of the `private_key` to
be stored in the HSM.

If only the `public_key` value is set an existing `private_key` kept
within the HSM can be reused (by design, it is not  possible to extract
`private_keys` from the HSM so if we only know the `public_key` from
logs then this is the behaviour we need).

If neither value is set then the existing functionality will be
preserved. For this reason the values are marked as both `Computed` and
`Optional` as this is appropriate per the terraform docs.